Separate SocketCache into file
